// InvoicerClient/src/app/api/model/invoicePaymentItem.ts
/**
 * Invoicer API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface InvoicePaymentItem {
  paymentId?: string;
  amount?: number;
  paidOn?: string;
  notes?: string | null;
}